arm/dt: hammerhead: change the VGH and VGL of DDIC
LCD has the after image due to increasing of pressure-resistant.
Because VGH/VGL of DDIC is set by +10V/-10V.
To prevent this issue, we need to chagne VGH/VGL of DDIC to +6V/-5V.
This is included in rev 1.0 Panel, But Previous HW doesn't have.
Therefore this is neccessary for those previous HW.

Change-Id: Ibca3550955a7e51b7a74d7e19609b56418299348
Signed-off-by: sooduek.kim <sooduek.kim@lge.com>
Signed-off-by: Devin Kim <dojip.kim@lge.com>